backend/requirements.txt
djangorestframework~=3.14.0
Django~=4.1.8
python-dotenv~=1.0.0
requests~=2.28.2
Pillow~=9.4.0
openai~=0.27.4
PyMonad~=2.4.0
gunicorn
gevent
django-cors-headers
djangorestframework-simplejwt
markdown
django-filter
black